+++ scripts/migration/src/index.js
@@ -1,7 +1,11 @@
 import { DockAPI } from '@docknetwork/dock-blockchain-api';
 import { Ed25519Keypair } from '@docknetwork/credential-sdk/keypairs';
 import { CheqdAPI, CheqdNetwork } from '@docknetwork/cheqd-blockchain-api';
-import { maybeToJSONString, timeout } from '@docknetwork/credential-sdk/utils';
+import {
+  maybeToJSONString,
+  timeout,
+  maybeToCheqdPayloadOrJSON,
+} from '@docknetwork/credential-sdk/utils';
 
 import pLimit from 'p-limit';
 import { checkBalance } from '@cheqd/sdk';
@@ -10,10 +14,16 @@ import Migration from './migration.js';
 
 const DOCK_ENDPOINT = process.env.DOCK_ENDPOINT || 'wss://mainnet-node.dock.io';
 const CHEQD_ENDPOINT = process.env.CHEQD_ENDPOINT || 'http://localhost:26657';
+const CHEQD_NETWORK = process.env.CHEQD_NETWORK || CheqdNetwork.Testnet;
 const ACCOUNT_COUNT = process.env.CHEQD_ACCOUNTS || 1;
-const BALANCE_SHARE = Number(process.env.BALANCE_SHARE || 1e6);
+const BALANCE_SHARE = Number(process.env.BALANCE_SHARE || 4e3);
 const MNEMONIC = process.env.CHEQD_MNEMONIC;
 const TRANSFER_FEE = Number(process.env.TRANSFER_FEE || 10000000);
+
+const { MASTER_KEY } = process.env;
+if (!MASTER_KEY) {
+  throw new Error('No master key');
+}
 
 const keyPairs = (process.env.SEEDS || '')
   .split(',')
@@ -24,14 +34,16 @@ const keyPairs = (process.env.SEEDS || '')
     Object.create(null),
   );
 
-const BREAK = Symbol('BREAK');
+keyPairs.MASTER_KEY = MASTER_KEY;
 
 const loopWithCatch = async (fn, catchFn) => {
+  const continueSym = Symbol('continue');
+
   for (;;) {
     try {
       return await fn();
     } catch (err) {
-      if (catchFn(err) === BREAK) break;
+      if (catchFn(err, continueSym) !== continueSym) break;
     }
   }
 };
@@ -40,7 +52,7 @@ const loopStreamWithCatch = async (stream, fn, catchFn) => {
   for await (const item of stream) {
     await loopWithCatch(
       () => fn(item),
-      (err) => catchFn(err, item),
+      (err, sym) => catchFn(err, item, sym),
     );
   }
 };
@@ -64,7 +76,7 @@ const transferNCHEQ = (from, to, ncheq) => {
 };
 
 const initCheqd = async (walletsCount = 0) => {
-  async function generateRandomWallet() {
+  async function generateRandomWallet(_, idx) {
     // Generate a random wallet
     const wallet = await DirectSecp256k1HdWallet.generate(24, {
       prefix: 'cheqd',
@@ -85,38 +97,50 @@ const initCheqd = async (walletsCount = 0) => {
       prefix: 'cheqd',
     })),
   );
-  const apis = await Promise.all(
-    wallets.map((wallet) => new CheqdAPI().init({
-      url: CHEQD_ENDPOINT,
-      network: CheqdNetwork.Testnet,
+
+  return await Promise.all(
+    wallets.map(async (wallet) => ({
+      api: await new CheqdAPI().init({
+        url: CHEQD_ENDPOINT,
+        network: CHEQD_NETWORK,
+        wallet,
+      }),
       wallet,
     })),
   );
-
-  return { wallets, apis };
 };
 
-async function shareBalance({ apis, wallets }) {
-  const [{ address: from }] = await wallets[0].getAccounts();
-  const initBalance = await getBalanceCheqd(from);
-  const txs = [];
+async function shareBalance(cheqds, spawn) {
+  const [{ address: from }] = await cheqds[0].wallet.getAccounts();
+  let initBalance = BigInt(BALANCE_SHARE * 1e9) * BigInt(Math.max(cheqds.length - 1, 0));
 
-  for (const wallet of wallets.slice(1)) {
-    const [{ address: to }] = await wallet.getAccounts();
+  const txs = [
+    ...(await Promise.all(
+      cheqds.slice(1).map(({ wallet }) => spawn(async () => {
+        const [{ address: to }] = await wallet.getAccounts();
+        try {
+          initBalance += await getBalanceCheqd(to);
+        } catch {}
 
-    txs.push(transferNCHEQ(from, to, BALANCE_SHARE * 1e9));
-  }
+        if (BALANCE_SHARE > 0) {
+          return transferNCHEQ(from, to, BALANCE_SHARE * 1e9);
+        } else {
+          return null;
+        }
+      })),
+    )),
+  ].filter(Boolean);
 
   const fee = {
     amount: [
-      { denom: 'ncheq', amount: String(TRANSFER_FEE * (wallets.length - 1)) },
+      { denom: 'ncheq', amount: String(TRANSFER_FEE * (cheqds.length - 1)) },
     ], // Fee amount in ncheq
-    gas: String(200000 * (wallets.length - 1)), // Gas limit
+    gas: String(200000 * (cheqds.length - 1)), // Gas limit
     payer: from,
   };
 
   if (txs.length) {
-    await apis[0].sdk.signer.signAndBroadcast(
+    await cheqds[0].api.sdk.signer.signAndBroadcast(
       from,
       txs,
       fee,
@@ -127,13 +151,19 @@ async function shareBalance({ apis, wallets }) {
   return initBalance;
 }
 
-async function payback({ apis, wallets }) {
-  const [{ address: to }] = await wallets[0].getAccounts();
+async function payback(cheqds) {
+  const [{ address: to }] = await cheqds[0].wallet.getAccounts();
+  let total = BigInt(0);
 
   await Promise.all(
-    wallets.slice(1).map(async (wallet, idx) => {
+    cheqds.slice(1).map(async ({ wallet, api }) => {
       const [{ address: from }] = await wallet.getAccounts();
       const balance = await getBalanceCheqd(from);
+      total += balance;
+
+      if (balance <= BigInt(TRANSFER_FEE)) {
+        return;
+      }
 
       const transfer = transferNCHEQ(from, to, balance - BigInt(TRANSFER_FEE));
 
@@ -143,7 +173,7 @@ async function payback({ apis, wallets }) {
         payer: from,
       };
 
-      await apis[idx + 1].sdk.signer.signAndBroadcast(
+      await api.sdk.signer.signAndBroadcast(
         from,
         [transfer],
         fee,
@@ -152,108 +182,149 @@ async function payback({ apis, wallets }) {
     }),
   );
 
-  return await getBalanceCheqd(to);
+  return total;
 }
 
-const getBalanceCheqd = async (account) => BigInt(
-  (await checkBalance(account, CHEQD_ENDPOINT)).find(
-    (balance) => balance.denom === 'ncheq',
-  ).amount,
-);
+const getBalanceCheqd = async (account) => {
+  const entries = await checkBalance(account, CHEQD_ENDPOINT);
+  const cheqd = entries.find((balance) => balance.denom === 'ncheq')?.amount;
+
+  return cheqd ? BigInt(cheqd) : BigInt(0);
+};
 
 const fmtNCHEQBalance = (ncheq) => `${(Number(ncheq) / 1e9).toFixed(9)} CHEQD`;
 
 async function main() {
+  const start = +new Date();
   const dock = new DockAPI();
+  const spawn = pLimit(10);
 
   await dock.init({ address: DOCK_ENDPOINT });
   dock.params = { address: DOCK_ENDPOINT };
 
   const cheqds = await initCheqd(+ACCOUNT_COUNT);
 
-  const initBalance = await shareBalance(cheqds);
-  console.log(
-    `Initial main account balance was ${fmtNCHEQBalance(initBalance)}`,
-  );
-
-  const spawn = pLimit(10);
-
-  const txs = new Migration(dock, cheqds.apis[0], keyPairs, spawn).txs();
+  const initBalance = await shareBalance(cheqds, spawn);
+  console.log(`Initial migration balance was ${fmtNCHEQBalance(initBalance)}`);
 
   let globalTxIdx = 0;
-  const loopSender = async (cheqd, idx) => {
-    let localTxIdx = 0;
-    const [{ address }] = await cheqd.sdk.options.wallet.getAccounts();
-
-    const sendTx = async (tx) => {
-      ++globalTxIdx;
-      ++localTxIdx;
-
-      console.log(
-        `Transaction #${globalTxIdx} (#${localTxIdx} from ${address}): ${maybeToJSONString(
-          tx,
-        )}`,
-      );
-
-      return await cheqd.signAndSend(tx);
-    };
-    const handleTxError = async (err, tx) => {
-      err.message = `Failed to execute transaction ${maybeToJSONString(tx)}: ${
-        err.message
-      }`;
-
-      if (err.message.includes('account sequence mismatch')) {
-      } else if (err.message.includes('fetch failed')) {
-        cheqd = await new CheqdAPI().init({
-          url: CHEQD_ENDPOINT,
-          network: CheqdNetwork.Testnet,
-          wallet: cheqds.wallets[idx],
-        });
-      } else if (
-        err.message.includes('DID Doc exists')
-        || err.message.includes('Resource exists')
-        || err.message.includes('tx already exists in cache')
-        || err.message.includes('DID Doc not found')
-        || err.message.includes('provided')
-      ) {
-        console.error(err);
-
-        return BREAK;
-      } else {
-        throw err;
-      }
-
-      console.error(err);
-    };
-
-    const handleDidTxs = (didTxs) => loopStreamWithCatch(didTxs, sendTx, handleTxError);
-    const handleDidTxsError = async (err) => {
-      err.message = `Failed to send transaction from ${
-        (await cheqds.wallets[idx].getAccounts())[0].address
-      }: \n${err.message}`;
-
-      if (
-        err.message.includes('No keypair')
-        || err.message.includes('Dock DID not found for')
-      ) {
-        console.error(err);
-      } else if (err.message.includes('fetch failed')) {
-        cheqd = await new CheqdAPI().init({
-          url: CHEQD_ENDPOINT,
-          network: CheqdNetwork.Testnet,
-          wallet: cheqds.wallets[idx],
-        });
-      } else {
-        throw err;
-      }
-    };
-
-    await loopStreamWithCatch(txs, handleDidTxs, handleDidTxsError);
-  };
-
+  let totalDLRs = 0;
+  let totalDIDs = 0;
   let err;
   try {
-    await Promise.all(cheqds.apis.map(loopSender));
+    const txs = new Migration(dock, cheqds[0].api, keyPairs, spawn).txs();
+
+    const loopSender = async ({ api, wallet }) => {
+      let localTxIdx = 0;
+      const [{ address }] = await api.sdk.options.wallet.getAccounts();
+
+      const reconnect = async () => {
+        console.log(`Reconnecting ${address}`);
+
+        await loopWithCatch(
+          async () => {
+            try {
+              await api.disconnect();
+            } finally {
+              await api.init({
+                url: CHEQD_ENDPOINT,
+                network: CHEQD_NETWORK,
+                wallet,
+              });
+            }
+          },
+          (err, continueSym) => {
+            console.error(err);
+
+            return continueSym;
+          },
+        );
+      };
+      const sendTx = async (tx) => {
+        console.log(
+          `Transaction #${
+            localTxIdx + 1
+          } sent from ${address}: ${maybeToJSONString(tx)}`,
+        );
+
+        const res = await api.signAndSend(tx);
+
+        ++globalTxIdx;
+        ++localTxIdx;
+        if (String(tx.typeUrl) === 'MsgCreateResource') {
+          totalDLRs++;
+        } else if (String(tx.typeUrl) === 'MsgCreateDidDoc') {
+          totalDIDs++;
+        }
+
+        console.log(
+          `Transaction #${localTxIdx} from ${address} was processed (#${globalTxIdx})`,
+        );
+
+        return res;
+      };
+      const handleTxError = async (err, tx, continueSym) => {
+        err.message = `Failed to execute transaction ${JSON.stringify(
+          maybeToCheqdPayloadOrJSON(tx),
+        )}: ${err.message}`;
+        const strErr = String(err);
+
+        if (
+          strErr.includes('account sequence mismatch')
+          || strErr.includes('incorrect account sequence')
+          || strErr.includes('is not initialized')
+        ) {
+        } else if (
+          strErr.includes('fetch failed')
+          || strErr.includes('Bad status')
+          || strErr.includes('other side closed')
+        ) {
+          await reconnect();
+        } else if (
+          strErr.includes('DID Doc exists')
+          || strErr.includes('Resource exists')
+          || strErr.includes('tx already exists in cache')
+          || strErr.includes('was submitted but was not yet found on the chain')
+          || strErr.includes('provided')
+        ) {
+          console.error(err);
+          console.log(`Skipping transaction #${localTxIdx} from ${address}`);
+
+          return;
+        } else {
+          throw err;
+        }
+
+        console.error(err);
+        return continueSym;
+      };
+
+      const handleDidTxs = (didTxs) => loopStreamWithCatch(didTxs(), sendTx, handleTxError);
+      const handleDidTxsError = async (err, continueSym) => {
+        err.message = `Failed to send transaction from ${
+          (await wallet.getAccounts())[0].address
+        }: \n${err.message}`;
+        const strErr = String(err);
+
+        if (
+          strErr.includes('No keypair')
+          || strErr.includes('Dock DID not found for')
+          || strErr.includes('the DID in question does not exist')
+        ) {
+          console.error(err);
+        } else if (strErr.includes('fetch failed')) {
+          await reconnect();
+        } else {
+          throw err;
+        }
+
+        return continueSym;
+      };
+
+      await loopStreamWithCatch(txs, handleDidTxs, handleDidTxsError);
+    };
+
+    await Promise.all(cheqds.slice(1).map(loopSender));
   } catch (error) {
     err = error;
   } finally {
@@ -262,22 +333,16 @@ async function main() {
       const endBalance = await payback(cheqds);
 
       console.log(
-        `Final main account balance is ${fmtNCHEQBalance(
-          endBalance,
-        )}. Totally spent ${fmtNCHEQBalance(
-          initBalance - endBalance,
-        )}. Average per sender is ${fmtNCHEQBalance(
-          (initBalance - endBalance) / (BigInt(ACCOUNT_COUNT) + BigInt(1)),
-        )}.`,
-      );
-
-      console.log(
-        `Total transactions sent: ${globalTxIdx}, average per sender: ${
-          globalTxIdx / (+ACCOUNT_COUNT + 1)
-        }`,
+        `Total spent: ${fmtNCHEQBalance(initBalance - endBalance)}
+Total time taken: ${new Date() - start} seconds
+Total transactions sent: ${globalTxIdx}, average per sender: ${
+  globalTxIdx / +ACCOUNT_COUNT
+}
+Created ${totalDIDs} DIDs and ${totalDLRs} DLRs`,
       );
     } catch (error) {
       error.message = `Payback failed: ${error.message}`;
+      console.error(error);
 
       err ||= error;
     }
@@ -290,7 +355,4 @@ async function main() {
 
 main()
   .then(() => process.exit(0))
-  .catch((error) => {
-    console.error(error);
-    process.exit(1);
-  });
+  .catch(console.error);