+++ packages/cheqd-blockchain-api/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @docknetwork/cheqd-blockchain-api
 
+## 0.18.0
+
+### Minor Changes
+
+- Align with the latest `cheqd` node version (`3.1.3`)
+
 ## 0.17.1
 
 ### Patch Changes
@@ -8,6 +14,8 @@
   - @docknetwork/credential-sdk@0.23.0
 
 ## 0.17.0
+
+### Minor Changes
 
 - Implementation of the `cheqd` migration and the associated identifier mappers.
 - Updated dependencies